feat: Organize view controllers
- Organize the roles of the reminder list view controller by extracting data source related behaviors into a separate file, `ReminderListViewController+DataSource.swift`.
	- Use the new data source to configure a list cell that displays the date and time of a reminder.
